+++ core/jvm/src/test/scala/fs2/MergeJoinSpec.scala
@@ -25,20 +25,20 @@ class MergeJoinSpec extends Fs2Spec {
 
     "merge/join consistency" in forAll { (s1: PureStream[Int], s2: PureStream[Int]) =>
       runLog { s1.get.through2v(s2.get.covary[IO])(pipe2.merge) }.toSet shouldBe
-      runLog { concurrent.join(2)(Stream(s1.get.covary[IO], s2.get.covary[IO])) }.toSet
+      runLog { Stream.join(2)(Stream(s1.get.covary[IO], s2.get.covary[IO])) }.toSet
     }
 
     "join (1)" in forAll { (s1: PureStream[Int]) =>
-      runLog { concurrent.join(1)(s1.get.covary[IO].map(Stream.emit)) }.toSet shouldBe runLog { s1.get }.toSet
+      runLog { Stream.join(1)(s1.get.covary[IO].map(Stream.emit)) }.toSet shouldBe runLog { s1.get }.toSet
     }
 
     "join (2)" in forAll { (s1: PureStream[Int], n: SmallPositive) =>
-      runLog { concurrent.join(n.get)(s1.get.covary[IO].map(Stream.emit)) }.toSet shouldBe
+      runLog { Stream.join(n.get)(s1.get.covary[IO].map(Stream.emit)) }.toSet shouldBe
       runLog { s1.get }.toSet
     }
 
     "join (3)" in forAll { (s1: PureStream[PureStream[Int]], n: SmallPositive) =>
-      runLog { concurrent.join(n.get)(s1.get.map(_.get.covary[IO]).covary[IO]) }.toSet shouldBe
+      runLog { Stream.join(n.get)(s1.get.map(_.get.covary[IO]).covary[IO]) }.toSet shouldBe
       runLog { s1.get.flatMap(_.get) }.toSet
     }
 
@@ -66,10 +66,10 @@ class MergeJoinSpec extends Fs2Spec {
       }
 
       "join" in {
-        runLog(concurrent.join(10)(Stream(full, hang)).take(1)) shouldBe Vector(42)
-        runLog(concurrent.join(10)(Stream(full, hang2)).take(1)) shouldBe Vector(42)
-        runLog(concurrent.join(10)(Stream(full, hang3)).take(1)) shouldBe Vector(42)
-        runLog(concurrent.join(10)(Stream(hang3,hang2,full)).take(1)) shouldBe Vector(42)
+        runLog(Stream.join(10)(Stream(full, hang)).take(1)) shouldBe Vector(42)
+        runLog(Stream.join(10)(Stream(full, hang2)).take(1)) shouldBe Vector(42)
+        runLog(Stream.join(10)(Stream(full, hang3)).take(1)) shouldBe Vector(42)
+        runLog(Stream.join(10)(Stream(hang3,hang2,full)).take(1)) shouldBe Vector(42)
       }
     }
   }
