+++ core/jvm/src/test/scala/fs2/StreamSpec.scala
@@ -75,7 +75,7 @@ class StreamSpec extends Fs2Spec with Inside {
     "onError (5)" in {
       val r = Stream.fail[IO](Err).onError(e => Stream.emit(e)).flatMap(Stream.emit).runLog.unsafeRunSync()
       val r2 = Stream.fail[IO](Err).onError(e => Stream.emit(e)).map(identity).runLog.unsafeRunSync()
-      val r3 = concurrent.join[IO, Int](4)(Stream(Stream.emit(1), Stream.fail(Err), Stream.emit(2)))
+      val r3 = Stream.join[IO, Int](4)(Stream(Stream.emit(1), Stream.fail(Err), Stream.emit(2)))
                          .attempt.runLog.unsafeRunSync()
       r shouldBe Vector(Err)
       r2 shouldBe Vector(Err)
